Change price type from float to double in Product

+++ Product.java
@@ -6,10 +6,10 @@ public class Product implements Recordable {
     private String manufacturerName;
     private String supplierName;
     private int quantity;
-    private float price;
+    private double price;
 
     public Product(String productID, String productName, String manufacturerName,
-                   String supplierName, int quantity, float price) {
+                   String supplierName, int quantity, double price) {
         this.productID = productID;
         this.productName = productName;
         this.manufacturerName = manufacturerName;
@@ -53,5 +53,5 @@ public class Product implements Recordable {
     public String getProductName() { return productName; }
     public String getManufacturerName() { return manufacturerName; }
     public String getSupplierName() { return supplierName; }
-    public float getPrice() { return price; }
+    public double getPrice() { return price; }
 }
